Warn user when attempting to delete non-existant order ID

+++ api.js
@@ -404,14 +404,25 @@ router.delete("/orders/products/:id",function(req,res){
 // Cancel order within 24 hrs. id is OrderID
 router.delete("/orders/:id",function(req,res){
 	authAndRun(req, res, function(req, res, customerID){
-		global.connection.query('DELETE FROM orders WHERE OrderID = ? AND OrderCustomerID = ? AND OrderDate >= NOW() - INTERVAL 1 DAY', [req.params.id, customerID], function (error, results, fields) {
+		global.connection.query('SELECT COUNT(*) AS OrderNum FROM orders WHERE OrderID = ?', [req.params.id], function (error, results, fields) {
 			if (error){
 				res.send(JSON.stringify({"status": 500, "error": error, "response": null}));
 				return;
 			}
-			global.connection.query('SELECT * FROM orders WHERE OrderID = ?', [req.params.id], function (error, results, fields) {
-				sendDeleteResult(res, error, results);
-			});
+			if (results.length > 0 && results[0]['OrderNum'] > 0) {
+				global.connection.query('DELETE FROM orders WHERE OrderID = ? AND OrderCustomerID = ? AND OrderDate >= NOW() - INTERVAL 1 DAY', [req.params.id, customerID], function (error, results, fields) {
+					if (error){
+						res.send(JSON.stringify({"status": 500, "error": error, "response": null}));
+						return;
+					}
+					global.connection.query('SELECT * FROM orders WHERE OrderID = ?', [req.params.id], function (error, results, fields) {
+						sendDeleteResult(res, error, results);
+					});
+				});
+			}
+			else {
+				res.send(JSON.stringify({"status": 500, "error": "Order ID doesn't exist", "response": null}));
+			}
 		});
 	});
 });
